Updated libraries and make call correctly

+++ app/src/main/java/com/simplekjl/showsapp/data/model/Show.kt
@@ -1,11 +1,13 @@
 package com.simplekjl.showsapp.data.model
 
 
-data class Show(val poster_path : String?,
-                val popularity : Int,
-                val id : Int,
-                val backdrop_path : String,
-                val vote_average : Int,
-                val overview : String,
-                val first_air_date : String,
-                val name : String)
+data class Show(
+    val poster_path: String?,
+    val popularity: Float,
+    val id: Int,
+    val backdrop_path: String,
+    val vote_average: Float,
+    val overview: String,
+    val first_air_date: String,
+    val name: String
+)
